+++ doc/source/changelog.rst
@@ -82,6 +82,12 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
   perturbable molecule with virtual sites. Each virtual site now inherits the REST2
   flag of its parent atom, so its charge is scaled along with it.
 
+* Fixed off-site charges (virtual sites) on non-ghost atoms of a perturbable molecule
+  being added to neither the ghost nor the non-ghost interaction group of the softcore
+  forces. Their interaction with any ghost atom was evaluated by the standard
+  ``NonbondedForce`` alone, i.e. without softening, and without the subtraction of the
+  coulomb energy that the softcore replaces.
+
 `2026.1.0 <https://github.com/openbiosim/sire/compare/2025.4.0...2026.1.0>`__ - June 2026
 -----------------------------------------------------------------------------------------